+++ hyperdbg/hprdbghv/code/hooks/ept-hook/EptHook.c
@@ -616,7 +616,7 @@ EptHookWriteAbsoluteJump2(PCHAR TargetBuffer, SIZE_T TargetAddress)
 }
 
 /**
- * @brief Hook ins
+ * @brief Hook instructions
  *
  * @param Hook The details of hooked pages
  * @param ProcessCr3 The target Process CR3
@@ -653,10 +653,11 @@ EptHookInstructionMemory(PEPT_HOOKED_PAGE_DETAIL Hook,
 
     //
     // Determine the number of instructions necessary to overwrite using Length Disassembler Engine
+    // EPTHOOK2 only supports 64-bit kernel (32-bit LDE is not supported)
     //
     for (SizeOfHookedInstructions = 0;
          SizeOfHookedInstructions < 19;
-         SizeOfHookedInstructions += ldisasm(((UINT64)TargetFunctionInSafeMemory + SizeOfHookedInstructions), TRUE))
+         SizeOfHookedInstructions += DisassemblerLengthDisassembleEngineInVmxRootOnTargetProcess((UINT64)TargetFunctionInSafeMemory + SizeOfHookedInstructions, FALSE))
     {
         //
         // Get the full size of instructions necessary to copy
@@ -1236,11 +1237,7 @@ EptHookHandleHookedPage(VIRTUAL_MACHINE_STATE *              VCpu,
     else if (!ViolationQualification.EptWriteable && ViolationQualification.WriteAccess)
     {
         //
-        // Test
-        //
-
-        //
-        // LogInfo("Guest RIP : 0x%llx tries to write on the page at :0x%llx", GuestRip, ExactAccessedAddress);
+        // LogInfo("Guest RIP : 0x%llx tries to write on the page at : 0x%llx", GuestRip, ExactAccessedAddress);
         //
 
         //
@@ -1251,11 +1248,7 @@ EptHookHandleHookedPage(VIRTUAL_MACHINE_STATE *              VCpu,
     else if (!ViolationQualification.EptReadable && ViolationQualification.ReadAccess)
     {
         //
-        // Test
-        //
-
-        //
-        // LogInfo("Guest RIP : 0x%llx tries to read the page at :0x%llx", GuestRip, ExactAccessedAddress);
+        // LogInfo("Guest RIP : 0x%llx tries to read the page at : 0x%llx", GuestRip, ExactAccessedAddress);
         //
 
         //