+++ hyperdbg/include/SDK/Headers/Constants.h
@@ -16,8 +16,8 @@
 //			 Version Information                //
 //////////////////////////////////////////////////
 
-#define VERSION_MAJOR 2
-#define VERSION_MINOR 2
+#define VERSION_MAJOR 0
+#define VERSION_MINOR 3
 #define VERSION_PATCH 0
 
 //